Move effect handlers and state files to dedicated packages
- Move all effect handlers from handler/ to effect/ package
- Move state-related files (StateExtensions, cache keys, etc.) to state/ package
- Update all imports across features and tests
- Clean structure: action/, state/, effect/, reducer/ subpackages
