+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/di/AppFeatureFactory.kt
@@ -4,7 +4,7 @@ import kotlinx.datetime.LocalDate
 import space.be1ski.vibits.shared.app.domain.model.AppState
 import space.be1ski.vibits.shared.app.presentation.action.AppAction
 import space.be1ski.vibits.shared.app.presentation.effect.AppEffect
-import space.be1ski.vibits.shared.app.presentation.handler.AppEffectHandler
+import space.be1ski.vibits.shared.app.presentation.effect.AppEffectHandler
 import space.be1ski.vibits.shared.app.presentation.reducer.appReducer
 import space.be1ski.vibits.shared.core.elm.Feature
 import space.be1ski.vibits.shared.core.elm.FeatureImpl
