+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/view/VibitsAppScaffold.kt
@@ -40,8 +40,8 @@ import space.be1ski.vibits.shared.feature.habits.domain.model.HabitsConfigEntry
 import space.be1ski.vibits.shared.feature.habits.domain.usecase.EarliestMemoDateUseCase
 import space.be1ski.vibits.shared.feature.habits.domain.usecase.IsActivityRangeBeforeUseCase
 import space.be1ski.vibits.shared.feature.habits.presentation.action.HabitsAction
-import space.be1ski.vibits.shared.feature.habits.presentation.getActivityData
 import space.be1ski.vibits.shared.feature.habits.presentation.state.HabitsState
+import space.be1ski.vibits.shared.feature.habits.presentation.state.getActivityData
 import space.be1ski.vibits.shared.feature.habits.presentation.view.components.rememberHabitsConfigTimeline
 import space.be1ski.vibits.shared.feature.memos.domain.model.Memo
 import space.be1ski.vibits.shared.feature.memos.presentation.state.MemosState
