+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/habits/di/HabitsFeatureFactory.kt
@@ -5,11 +5,11 @@ import space.be1ski.vibits.shared.core.elm.FeatureImpl
 import space.be1ski.vibits.shared.feature.habits.domain.usecase.CalculateActivityDataUseCase
 import space.be1ski.vibits.shared.feature.habits.domain.usecase.PrewarmActivityDataUseCase
 import space.be1ski.vibits.shared.feature.habits.presentation.action.HabitsAction
+import space.be1ski.vibits.shared.feature.habits.presentation.effect.HabitsActivityEffectHandler
 import space.be1ski.vibits.shared.feature.habits.presentation.effect.HabitsEffect
-import space.be1ski.vibits.shared.feature.habits.presentation.handler.HabitsActivityEffectHandler
-import space.be1ski.vibits.shared.feature.habits.presentation.handler.HabitsEffectHandler
-import space.be1ski.vibits.shared.feature.habits.presentation.handler.HabitsMemoEffectHandler
-import space.be1ski.vibits.shared.feature.habits.presentation.handler.HabitsRefreshEffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.effect.HabitsEffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.effect.HabitsMemoEffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.effect.HabitsRefreshEffectHandler
 import space.be1ski.vibits.shared.feature.habits.presentation.reducer.habitsReducer
 import space.be1ski.vibits.shared.feature.habits.presentation.state.HabitsState
