+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/habits/presentation/reducer/HabitsReducer.kt
@@ -2,14 +2,6 @@ package space.be1ski.vibits.shared.feature.habits.presentation.reducer
 import space.be1ski.vibits.shared.core.elm.Reducer
 import space.be1ski.vibits.shared.feature.habits.presentation.action.HabitsAction
 import space.be1ski.vibits.shared.feature.habits.presentation.effect.HabitsEffect
-import space.be1ski.vibits.shared.feature.habits.presentation.reducer.cacheReducer
-import space.be1ski.vibits.shared.feature.habits.presentation.reducer.configDeleteReducer
-import space.be1ski.vibits.shared.feature.habits.presentation.reducer.configReducer
-import space.be1ski.vibits.shared.feature.habits.presentation.reducer.configWarningReducer
-import space.be1ski.vibits.shared.feature.habits.presentation.reducer.editorReducer
-import space.be1ski.vibits.shared.feature.habits.presentation.reducer.responseReducer
-import space.be1ski.vibits.shared.feature.habits.presentation.reducer.selectionReducer
-import space.be1ski.vibits.shared.feature.habits.presentation.reducer.singleToggleReducer
 import space.be1ski.vibits.shared.feature.habits.presentation.state.HabitsState
 
 /**
