+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/habits/presentation/state/HabitsState.kt
@@ -7,8 +7,8 @@ import space.be1ski.vibits.shared.feature.habits.domain.model.ContributionDay
 import space.be1ski.vibits.shared.feature.habits.domain.model.DailyMemoInfo
 import space.be1ski.vibits.shared.feature.habits.domain.model.HabitConfig
 import space.be1ski.vibits.shared.feature.habits.domain.normalizeHabitTag
-import space.be1ski.vibits.shared.feature.habits.presentation.ActivityCacheKey
-import space.be1ski.vibits.shared.feature.habits.presentation.CachedActivityData
+import space.be1ski.vibits.shared.feature.habits.presentation.state.ActivityCacheKey
+import space.be1ski.vibits.shared.feature.habits.presentation.state.CachedActivityData
 import space.be1ski.vibits.shared.feature.memos.domain.model.Memo
 
 /**
