+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/onboarding/di/OnboardingFeatureFactory.kt
@@ -3,11 +3,11 @@ package space.be1ski.vibits.shared.feature.onboarding.di
 import space.be1ski.vibits.shared.core.elm.Feature
 import space.be1ski.vibits.shared.core.elm.FeatureImpl
 import space.be1ski.vibits.shared.feature.onboarding.presentation.action.OnboardingAction
+import space.be1ski.vibits.shared.feature.onboarding.presentation.effect.OnboardingCompletionEffectHandler
 import space.be1ski.vibits.shared.feature.onboarding.presentation.effect.OnboardingEffect
-import space.be1ski.vibits.shared.feature.onboarding.presentation.handler.OnboardingCompletionEffectHandler
-import space.be1ski.vibits.shared.feature.onboarding.presentation.handler.OnboardingEffectHandler
-import space.be1ski.vibits.shared.feature.onboarding.presentation.handler.OnboardingPresetsEffectHandler
-import space.be1ski.vibits.shared.feature.onboarding.presentation.handler.OnboardingSetupEffectHandler
+import space.be1ski.vibits.shared.feature.onboarding.presentation.effect.OnboardingEffectHandler
+import space.be1ski.vibits.shared.feature.onboarding.presentation.effect.OnboardingPresetsEffectHandler
+import space.be1ski.vibits.shared.feature.onboarding.presentation.effect.OnboardingSetupEffectHandler
 import space.be1ski.vibits.shared.feature.onboarding.presentation.reducer.onboardingReducer
 import space.be1ski.vibits.shared.feature.onboarding.presentation.state.OnboardingState
