+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/onboarding/presentation/effect/OnboardingCompletionEffectHandler.kt
@@ -1,4 +1,4 @@
-package space.be1ski.vibits.shared.feature.onboarding.presentation.handler
+package space.be1ski.vibits.shared.feature.onboarding.presentation.effect
 
 import kotlinx.coroutines.flow.Flow
 import space.be1ski.vibits.shared.core.elm.EffectHandler
