+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/settings/di/SettingsFeatureFactory.kt
@@ -5,11 +5,11 @@ import space.be1ski.vibits.shared.core.elm.Feature
 import space.be1ski.vibits.shared.core.elm.FeatureImpl
 import space.be1ski.vibits.shared.feature.mode.domain.model.AppMode
 import space.be1ski.vibits.shared.feature.settings.presentation.action.SettingsAction
+import space.be1ski.vibits.shared.feature.settings.presentation.effect.SettingsCredentialsEffectHandler
 import space.be1ski.vibits.shared.feature.settings.presentation.effect.SettingsEffect
-import space.be1ski.vibits.shared.feature.settings.presentation.handler.SettingsCredentialsEffectHandler
-import space.be1ski.vibits.shared.feature.settings.presentation.handler.SettingsEffectHandler
-import space.be1ski.vibits.shared.feature.settings.presentation.handler.SettingsModeEffectHandler
-import space.be1ski.vibits.shared.feature.settings.presentation.handler.SettingsPreferencesEffectHandler
+import space.be1ski.vibits.shared.feature.settings.presentation.effect.SettingsEffectHandler
+import space.be1ski.vibits.shared.feature.settings.presentation.effect.SettingsModeEffectHandler
+import space.be1ski.vibits.shared.feature.settings.presentation.effect.SettingsPreferencesEffectHandler
 import space.be1ski.vibits.shared.feature.settings.presentation.reducer.settingsReducer
 import space.be1ski.vibits.shared.feature.settings.presentation.state.SettingsState
