+++ shared/src/commonTest/kotlin/space/be1ski/vibits/shared/feature/habits/presentation/HabitsCacheTest.kt
@@ -8,6 +8,8 @@ import space.be1ski.vibits.shared.feature.habits.domain.model.ActivityWeekData
 import space.be1ski.vibits.shared.feature.habits.presentation.action.HabitsAction
 import space.be1ski.vibits.shared.feature.habits.presentation.effect.HabitsEffect
 import space.be1ski.vibits.shared.feature.habits.presentation.reducer.habitsReducer
+import space.be1ski.vibits.shared.feature.habits.presentation.state.ActivityCacheKey
+import space.be1ski.vibits.shared.feature.habits.presentation.state.CachedActivityData
 import space.be1ski.vibits.shared.feature.habits.presentation.state.HabitsState
 import space.be1ski.vibits.shared.feature.mode.domain.model.AppMode
 import kotlin.test.Test
