+++ shared/src/commonTest/kotlin/space/be1ski/vibits/shared/feature/habits/presentation/HabitsEffectHandlerTest.kt
@@ -5,11 +5,11 @@ import kotlinx.datetime.LocalDate
 import space.be1ski.vibits.shared.app.domain.model.ActivityMode
 import space.be1ski.vibits.shared.app.domain.model.ActivityRange
 import space.be1ski.vibits.shared.feature.habits.presentation.action.HabitsAction
+import space.be1ski.vibits.shared.feature.habits.presentation.effect.HabitsActivityEffectHandler
 import space.be1ski.vibits.shared.feature.habits.presentation.effect.HabitsEffect
-import space.be1ski.vibits.shared.feature.habits.presentation.handler.HabitsActivityEffectHandler
-import space.be1ski.vibits.shared.feature.habits.presentation.handler.HabitsEffectHandler
-import space.be1ski.vibits.shared.feature.habits.presentation.handler.HabitsMemoEffectHandler
-import space.be1ski.vibits.shared.feature.habits.presentation.handler.HabitsRefreshEffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.effect.HabitsEffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.effect.HabitsMemoEffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.effect.HabitsRefreshEffectHandler
 import space.be1ski.vibits.shared.feature.habits.presentation.reducer.habitsReducer
 import space.be1ski.vibits.shared.feature.memos.domain.model.Memo
 import space.be1ski.vibits.shared.feature.mode.domain.model.AppMode
