+++ shared/src/commonTest/kotlin/space/be1ski/vibits/shared/feature/habits/presentation/HabitsStateExtensionsTest.kt
@@ -4,7 +4,11 @@ import space.be1ski.vibits.shared.app.domain.model.ActivityMode
 import space.be1ski.vibits.shared.app.domain.model.ActivityRange
 import space.be1ski.vibits.shared.feature.habits.domain.model.ActivityWeekData
 import space.be1ski.vibits.shared.feature.habits.presentation.reducer.habitsReducer
+import space.be1ski.vibits.shared.feature.habits.presentation.state.ActivityCacheKey
+import space.be1ski.vibits.shared.feature.habits.presentation.state.CachedActivityData
 import space.be1ski.vibits.shared.feature.habits.presentation.state.HabitsState
+import space.be1ski.vibits.shared.feature.habits.presentation.state.getActivityData
+import space.be1ski.vibits.shared.feature.habits.presentation.state.isDataLoading
 import space.be1ski.vibits.shared.feature.mode.domain.model.AppMode
 import kotlin.test.Test
 import kotlin.test.assertEquals
