+++ shared/src/commonTest/kotlin/space/be1ski/vibits/shared/feature/memos/presentation/MemosEffectHandlerTest.kt
@@ -11,11 +11,11 @@ import space.be1ski.vibits.shared.feature.memos.domain.usecase.LoadCachedMemosUs
 import space.be1ski.vibits.shared.feature.memos.domain.usecase.LoadMemosUseCase
 import space.be1ski.vibits.shared.feature.memos.domain.usecase.UpdateMemoUseCase
 import space.be1ski.vibits.shared.feature.memos.presentation.action.MemosAction
+import space.be1ski.vibits.shared.feature.memos.presentation.effect.MemosCredentialsEffectHandler
 import space.be1ski.vibits.shared.feature.memos.presentation.effect.MemosEffect
-import space.be1ski.vibits.shared.feature.memos.presentation.handler.MemosCredentialsEffectHandler
-import space.be1ski.vibits.shared.feature.memos.presentation.handler.MemosEffectHandler
-import space.be1ski.vibits.shared.feature.memos.presentation.handler.MemosLoadEffectHandler
-import space.be1ski.vibits.shared.feature.memos.presentation.handler.MemosWriteEffectHandler
+import space.be1ski.vibits.shared.feature.memos.presentation.effect.MemosEffectHandler
+import space.be1ski.vibits.shared.feature.memos.presentation.effect.MemosLoadEffectHandler
+import space.be1ski.vibits.shared.feature.memos.presentation.effect.MemosWriteEffectHandler
 import space.be1ski.vibits.shared.feature.memos.presentation.reducer.memosReducer
 import space.be1ski.vibits.shared.test.FakeCredentialsRepository
 import space.be1ski.vibits.shared.test.FakeMemosRepository
